Update CUDA kernels and clean codes.
