+++ deepspeed/inference/v2/kernels/core_ops/core_ops.cpp
@@ -39,5 +39,4 @@ PYBIND11_MODULE(TORCH_EXTENSION_NAME, m)
     m.def("cuda_wf6af16_linear", &cuda_wf6af16_linear, "DeepSpeed Wf6Af16 linear in CUDA");
     m.def(
         "preprocess_weight", &preprocess_weight, "preprocess the FP16 weight to be 2bit and 4 bit");
-    m.def("preprocess_scales", &preprocess_scales, "preprocess the FP16 scales");
 }
