+++ op_builder/inference_core_ops.py
@@ -67,10 +67,8 @@ class InferenceCoreBuilder(CUDAOpBuilder):
             "inference/v2/kernels/core_ops/cuda_rms_norm/rms_norm_cuda.cu",
             "inference/v2/kernels/core_ops/gated_activations/gated_activation_kernels.cpp",
             "inference/v2/kernels/core_ops/gated_activations/gated_activation_kernels_cuda.cu",
-            "inference/v2/kernels/core_ops/cuda_linear/cuda_linear_kernels.cu",
             "inference/v2/kernels/core_ops/cuda_linear/cuda_linear_kernels.cpp",
             "inference/v2/kernels/core_ops/cuda_linear/Launcher.cu",
-            "inference/v2/kernels/core_ops/cuda_linear/GenMatrix_QuantLLM.cpp",
         ]
 
         prefix = self.get_prefix()
